requirements.txt
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
peewee==3.14.4
psycopg2-binary==2.8.6
pycparser==2.20
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
